remove redundant files and SCSS
This removes a bunch of custom styles that weren't actually being used by the docs site. Docs are configured to be served from the root, so nothing in pages is actually used and we can safely remove it.

The custom.css file has been updated to matchthe Hackney developer playbook